refactor: package name redis->cache fix: go mod rest->redis
